refactor: use native es modules
