+++ audits/index.html
@@ -12,7 +12,7 @@
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;600&family=JetBrains+Mono&display=swap" rel="stylesheet">
   <link rel="stylesheet" href="styles.css">
-  <script src="/scripts/viewer.js?v=3" defer></script>
+  <script type="module" src="./main.js"></script>
 </head>
   <body data-theme="dark">
   <header id="topBar" class="top-bar">
