+++ audits/main.js
@@ -1,3 +1,3 @@
-import { init } from './modules/audits.js';
+import { init } from './audits.js';
 
 document.addEventListener('DOMContentLoaded', init);
