+++ package.json
@@ -3,13 +3,11 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "build": "esbuild audits/scripts/main.js --bundle --outfile=audits/scripts/viewer.js",
-    "lint": "eslint \"audits/scripts/**/*.js\"",
-    "format": "prettier -w \"audits/scripts/**/*.js\"",
+    "lint": "eslint \"audits/**/*.js\"",
+    "format": "prettier -w \"audits/**/*.js\"",
     "test": "./tests/run.sh"
   },
   "devDependencies": {
-    "esbuild": "^0.19.0",
     "eslint": "^8.57.0",
     "prettier": "^3.2.5"
   }
